+++ app/controllers/concerns/shopify_app/authenticated.rb
@@ -8,7 +8,7 @@ module ShopifyApp
       include ShopifyApp::Localization
       include ShopifyApp::LoginProtection
       include ShopifyApp::EmbeddedApp
-      before_action :login_again_if_different_shop
+      before_action :login_again_if_different_user_or_shop
       around_action :shopify_session
     end
   end